+++ data-raw/adeg.R
@@ -41,12 +41,12 @@ gen_adeg <- function(seed = 123) {
     )),
     PARAM = as.factor(dplyr::case_when(
       PARAMCD == "EGHRMN" ~ "ECG Mean Heart Rate (beats/min)",
-      PARAMCD == "PRAG" ~ "PR Interval, Aggregate (msec)",
-      PARAMCD == "QRSAG" ~ "QRS Duration, Aggregate (msec)",
-      PARAMCD == "QTC" ~ "QT Interval, Corrected (msec)",
-      PARAMCD == "QTCBAG" ~ "QTcB Interval, Aggregate (msec)",
-      PARAMCD == "QTCFAG" ~ "QTcF Interval, Aggregate (msec)",
-      PARAMCD == "RRAG" ~ "RR Interval, Aggregate (msec)",
+      PARAMCD == "PRAG" ~ "PR Interval, Aggregate (ms)",
+      PARAMCD == "QRSAG" ~ "QRS Duration, Aggregate (ms)",
+      PARAMCD == "QTC" ~ "QT Interval, Corrected (ms)",
+      PARAMCD == "QTCBAG" ~ "QTcB Interval, Aggregate (ms)",
+      PARAMCD == "QTCFAG" ~ "QTcF Interval, Aggregate (ms)",
+      PARAMCD == "RRAG" ~ "RR Interval, Aggregate (ms)",
       PARAMCD == "INTP" ~ "Interpretation"
     )),
     AVISIT = forcats::fct_reorder(
@@ -118,20 +118,20 @@ gen_adeg <- function(seed = 123) {
       .default = NA
     ),
     CRIT1 = dplyr::case_when(
-      PARAMCD == "EGHRMN" ~ "<50",
-      PARAMCD == "PRAG" ~ "<120"
-    ),
-    CRIT1FL = dplyr::case_when(
-      PARAMCD == "EGHRMN" & AVAL < 50 ~ "Y",
-      PARAMCD == "PRAG" & AVAL < 120 ~ "Y"
-    ),
-    CRIT2 = dplyr::case_when(
       PARAMCD == "EGHRMN" ~ ">100",
       PARAMCD == "PRAG" ~ ">200"
     ),
-    CRIT2FL = dplyr::case_when(
+    CRIT1FL = dplyr::case_when(
       PARAMCD == "EGHRMN" & AVAL > 100 ~ "Y",
       PARAMCD == "PRAG" & AVAL > 200 ~ "Y"
+    ),
+    CRIT2 = dplyr::case_when(
+      PARAMCD == "EGHRMN" ~ "<50",
+      PARAMCD == "PRAG" ~ "<120"
+    ),
+    CRIT2FL = dplyr::case_when(
+      PARAMCD == "EGHRMN" & AVAL < 50 ~ "Y",
+      PARAMCD == "PRAG" & AVAL < 120 ~ "Y"
     ),
     EGCLSIG = as.factor(dplyr::case_when(
       AVAL > 500 & (PARAMCD == "QTC" | PARAMCD == "QTCBAG" | PARAMCD == "QTCFAG") ~ "Y",
@@ -145,11 +145,12 @@ gen_adeg <- function(seed = 123) {
   # and set EGCLSIG accordingly
   intp_rows <- which(gen$PARAMCD == "INTP")
   # Randomly assign NORMAL or ABNORMAL to AVALC for INTP rows
-  gen$AVALC[intp_rows] <- sample(c("NORMAL", "ABNORMAL"), length(intp_rows), replace = TRUE, prob = c(0.5, 0.5))
+  gen$AVALC[intp_rows] <- sample(c("ABNORMAL, CS", "ABNORMAL, NCS", "NORMAL"),
+                                 length(intp_rows), replace = TRUE, prob = c(0.5, 0.5, 0.5))
 
   # Set EGCLSIG based on AVALC
   normal_rows <- intp_rows[gen$AVALC[intp_rows] == "NORMAL"]
-  abnormal_rows <- intp_rows[gen$AVALC[intp_rows] == "ABNORMAL"]
+  abnormal_rows <- intp_rows[gen$AVALC[intp_rows] == "ABNORMAL, CS"]
 
   # Set EGCLSIG to NA for NORMAL rows
   gen$EGCLSIG[normal_rows] <- NA
@@ -235,7 +236,8 @@ gen_adeg <- function(seed = 123) {
     "SAFFL",
     "STUDYID",
     "AGE",
-    "SEX"
+    "SEX",
+    "RACE"
   )
 
   # Select only the key and the 'to_keep' variables from ADSL
@@ -269,6 +271,12 @@ gen_adeg <- function(seed = 123) {
 
     # Categorization variable
     BASECAT1 = "Baseline Category 1"
+  )
+
+  gen <- dplyr::mutate(
+    gen,
+    CRIT1 = factor(CRIT1, levels = c(">200", ">100")),
+    CRIT2 = factor(CRIT2, levels = c("<50", "<120"))
   )
 
   # Handle NA values and convert characters to factors